+++ packages/paste-libraries/clipboard-copy/build.js
