+++ packages/paste-libraries/clipboard-copy/src/index.tsx
