+++ packages/paste-libraries/clipboard-copy/tsconfig.json
